refactor (RateLimiter) add LeagueJSRateLimiter * wraps RateLimiter instances for every region * provides more convenient update and scheduling methods
