+++ lib/Endpoint.js
@@ -43,7 +43,7 @@ class Endpoint {
 
 	}
 
-	getConfig(){
+	getConfig() {
 		return this.config;
 	}
 
@@ -121,14 +121,11 @@ class Endpoint {
 				if (cacheValue) {
 					return cacheValue;
 				} else {
-					return this.rateLimiter.per10[platformId].scheduling(
-						() => this.rateLimiter.per600[platformId].scheduling(
-							() => ApiRequest.executing(requestUrl, this.requestOptions)
-						)
-					).then(response => {
-						if (this.cache) {this.cache.set(requestUrl, response);}
-						return response;
-					});
+					return this.rateLimiter.scheduling(platformId, () => ApiRequest.executing(requestUrl, this.requestOptions))
+						.then(response => {
+							if (this.cache) {this.cache.set(requestUrl, response);}
+							return response;
+						});
 				}
 			});
 	}
