+++ lib/LeagueJS.js
@@ -5,7 +5,7 @@
 // const compatibility = require('./compatibility');
 
 const Config = require('./Config.js');
-
+const LeagueJSRateLimiter = require('./ratelimiter/LeagueJSRateLimiter');
 // Endpoints
 
 const SummonerEndpoint = require('./endpoints/SummonerEndpoint');
@@ -20,7 +20,6 @@ const SpectatorEndpoint = require('./endpoints/SpectatorEndpoint');
 const StaticDataEndpoint = require('./endpoints/StaticDataEndpoint');
 
 
-const LeagueUtil = require('./LeagueUtil');
 const util = require('./util');
 
 const DEFAULT_PLATFORM_ID = 'na1';
@@ -56,7 +55,7 @@ class LeagueJS {
 
 		// rate.setLimits = this.config.limits; // TODO: add when reworking RateLimiter
 		let {per10, per600, allowBursts} = this.config.limits;
-		this.rateLimiter = LeagueUtil.createRateLimiter(per10, per600, allowBursts);
+		this.rateLimiter = new LeagueJSRateLimiter(per10, per600, allowBursts);
 
 		// setting the Endpoints explicitly for better code completion and IDE support
 		// Endpoints that take a rate-limiter as constructor parameter are rate-limited.
@@ -151,7 +150,7 @@ class LeagueJS {
 	 * @param allowBursts
 	 */
 	setRateLimit(per10 = this.config.limits.per10, per600 = this.config.limits.per10, allowBursts = this.config.limits.allowBursts){
-		this.rateLimiter = LeagueUtil.updateRateLimiter(this.rateLimiter, per10, per600, allowBursts);
+		this.rateLimiter.update(per10, per600, allowBursts);
 	}
 	getRateLimits(){
 		return this.rateLimiter.limits;
