+++ lib/util/MatchUtil.js
@@ -15,9 +15,9 @@ class MatchUtil {
 		});
 
 		if (ddVersion === null) {
-			throw new Error(`LeagueUtil.getDdragonVersionForMatchVersion(): no ddragon Version found from gameVersion: ${gameVersionMajorMinor}`);
+			throw new Error(`MatchUtil.getVersionForGameVersion(): no ddragon Version found from gameVersion: ${gameVersionMajorMinor}`);
 		} else {
-			console.log(`LeagueUtil.getDdragonVersionForMatchVersion() version is resolved: ${ddVersion} gameVersion: ${gameVersion} majorMinor: ${gameVersionMajorMinor}`);
+			console.log(`MatchUtil.getVersionForGameVersion() version is resolved: ${ddVersion} gameVersion: ${gameVersion} majorMinor: ${gameVersionMajorMinor}`);
 			return ddVersion;
 		}
 	}
