+++ test/LeagueJS.test.js
@@ -21,7 +21,8 @@ describe('League of Legends api wrapper test suite', function() {
 			new LeagueJs();
 		}, 'Did not throw error with empty API_KEY.').to.throw('apiKey');
 	});
-	it('inherits all the Utility functionality from LeagueUtil', function () { // TODO: rework
+	it('provides all Utility classes', function () {
+		// TODO: add Test if all Util classes were added!?
 		LeagueJs.util.should.have.property('EndpointUtil');
 	});
